refactor: Replace PersonalityCard with SettingsCard
