+++ apps/desktop/src/components/settings/personalities/index.ts
@@ -1,6 +1,5 @@
 export { PersonalitiesSection } from "./PersonalitiesSection";
 export { PersonalitiesList } from "./PersonalitiesList";
-export { PersonalityCard } from "./PersonalityCard";
 export { PersonalityForm } from "./PersonalityForm";
 export { PersonalitySlider } from "./PersonalitySlider";
 export { DynamicBehaviorSections } from "./DynamicBehaviorSections";
